task-platform/task-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
requests
cryptography>=38.0.0
PyJWT>=2.8.0
